# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: RentaCar Operational Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /fleet:
    get:
      operationId: listFleet
      tags: [fleet]
      summary: List all vehicles
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [disponivel, alugado, manutencao, reservado_temporario]
        - name: categoria
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of vehicles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Vehicle"
    post:
      operationId: createVehicle
      tags: [fleet]
      summary: Add a new vehicle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VehicleInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"
  /fleet/{id}:
    get:
      operationId: getVehicle
      tags: [fleet]
      summary: Get a vehicle by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Vehicle details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"
        "404":
          description: Not found
    patch:
      operationId: updateVehicle
      tags: [fleet]
      summary: Update a vehicle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VehicleUpdate"
      responses:
        "200":
          description: Updated vehicle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"
    delete:
      operationId: deleteVehicle
      tags: [fleet]
      summary: Delete a vehicle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /fleet/{id}/toggle-status:
    patch:
      operationId: toggleVehicleStatus
      tags: [fleet]
      summary: Toggle vehicle availability/maintenance status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VehicleStatusToggle"
      responses:
        "200":
          description: Updated vehicle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"
  /fleet/{id}/damage-history:
    get:
      operationId: getVehicleDamageHistory
      tags: [fleet]
      summary: Get damage photo history for a vehicle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Damage history entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DamageEntry"
...
components:
  schemas:
...
    Vehicle:
      type: object
      required: [id, marca_modelo, categoria, preco_base_dia, valor_caucao, extra_franquia_zero, status]
      properties:
        id:
          type: integer
        marca_modelo:
          type: string
        categoria:
          type: string
          enum: [Economico, Familiar, SUV, Premium]
        foto_url:
          type: ["string", "null"]
        preco_base_dia:
          type: number
        valor_caucao:
          type: number
        extra_franquia_zero:
          type: number
        status:
          type: string
          enum: [disponivel, alugado, manutencao, reservado_temporario]
        created_at:
          type: ["string", "null"]
    VehicleInput:
      type: object
      required: [marca_modelo, categoria, preco_base_dia, valor_caucao, extra_franquia_zero]
      properties:
        marca_modelo:
          type: string
        categoria:
          type: string
          enum: [Economico, Familiar, SUV, Premium]
        foto_url:
          type: string
        preco_base_dia:
          type: number
        valor_caucao:
          type: number
        extra_franquia_zero:
          type: number
        status:
          type: string
          enum: [disponivel, alugado, manutencao, reservado_temporario]
    VehicleUpdate:
      type: object
      properties:
        marca_modelo:
          type: string
        categoria:
          type: string
          enum: [Economico, Familiar, SUV, Premium]
        foto_url:
          type: string
        preco_base_dia:
          type: number
        valor_caucao:
          type: number
        extra_franquia_zero:
          type: number
        status:
          type: string
          enum: [disponivel, alugado, manutencao, reservado_temporario]
    VehicleStatusToggle:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [disponivel, manutencao]
    DamageEntry:
      type: object
      required: [reservation_id, cliente_nome, fotos, created_at]
      properties:
        reservation_id:
          type: integer
        cliente_nome:
          type: string
        fotos:
          type: array
          items:
            type: string
        created_at:
          type: string
...