# lib/api-zod/src/generated/types/damageEntry.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * RentaCar Operational Management API
 * OpenAPI spec version: 0.1.0
 */

export interface DamageEntry {
  reservation_id: number;
  cliente_nome: string;
  fotos: string[];
  created_at: string;
}