# lib/api-zod/src/generated/types/vehicle.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * RentaCar Operational Management API
 * OpenAPI spec version: 0.1.0
 */
import type { VehicleCategoria } from './vehicleCategoria';
import type { VehicleStatus } from './vehicleStatus';

export interface Vehicle {
  id: number;
  marca_modelo: string;
  categoria: VehicleCategoria;
  /** @nullable */
  foto_url?: string | null;
  preco_base_dia: number;
  valor_caucao: number;
  extra_franquia_zero: number;
  status: VehicleStatus;
  /** @nullable */
  created_at?: string | null;
}