# lib/api-zod/src/generated/types/vehicleInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * RentaCar Operational Management API
 * OpenAPI spec version: 0.1.0
 */
import type { VehicleInputCategoria } from './vehicleInputCategoria';
import type { VehicleInputStatus } from './vehicleInputStatus';

export interface VehicleInput {
  marca_modelo: string;
  categoria: VehicleInputCategoria;
  foto_url?: string;
  preco_base_dia: number;
  valor_caucao: number;
  extra_franquia_zero: number;
  status?: VehicleInputStatus;
}