# lib/api-zod/src/generated/types/vehicleUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * RentaCar Operational Management API
 * OpenAPI spec version: 0.1.0
 */
import type { VehicleUpdateCategoria } from './vehicleUpdateCategoria';
import type { VehicleUpdateStatus } from './vehicleUpdateStatus';

export interface VehicleUpdate {
  marca_modelo?: string;
  categoria?: VehicleUpdateCategoria;
  foto_url?: string;
  preco_base_dia?: number;
  valor_caucao?: number;
  extra_franquia_zero?: number;
  status?: VehicleUpdateStatus;
}